# be/docs/swagger.yaml
basePath: /
definitions:
...
  api.ErrorResponse:
    properties:
      code:
        description: Error code for frontend i18n
        type: string
      message:
        description: Fallback message in English
        type: string
      params:
        additionalProperties:
          type: string
        description: Dynamic parameters for interpolation
        type: object
    type: object
...
host: localhost:1971
info:
  contact:
    email: support@example.com
    name: API Support
  description: API documentation for R-PRJ-NG backend
  license:
    name: Apache 2.0
    url: https://opensource.org/licenses/Apache-2.0
  termsOfService: http://swagger.io/terms/
  title: ρBee (rhobee) API
  version: "1.0"
paths:
  /countries:
    get:
      description: Returns a list of all countries from the countrylist table
      produces:
      - application/json
      responses:
        "200":
          description: List of countries
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: returns all countries from countrylist table
      tags:
      - navigation
  /countries/{countryId}:
    get:
      description: Returns a country from the countrylist table by its ID
      parameters:
      - description: Country ID
        in: path
        name: countryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Country data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Country not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: returns a country from countrylist table
      tags:
      - navigation
...
  /nav/{objectId}:
    get:
      description: Returns the navigation object specified by its ID
      parameters:
      - description: Object ID
        in: path
        name: objectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Navigation object data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Object not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: returns a navigation object by its ID
      tags:
      - navigation
  /nav/{objectId}/indexes:
    get:
      description: Returns index pages located directly under the specified object
        ID
      parameters:
      - description: Object ID
        in: path
        name: objectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of index pages
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Object not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: returns index pages under a given object
      tags:
      - navigation
  /nav/breadcrumb/{objectId}:
    get:
      description: Returns the breadcrumb trail for the specified object ID
      parameters:
      - description: Object ID
        in: path
        name: objectId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Breadcrumb data
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Object not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: returns breadcrumb for a given object
      tags:
      - navigation
  /nav/children/{folderId}:
    get:
      description: Returns the list of child objects under the specified folder ID
      parameters:
      - description: Folder ID
        in: path
        name: folderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of child objects
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Folder not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: returns children of a given folder
      tags:
      - navigation
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...